requirements.txt
# !-- Make sure to use Python 3.10 and not a higher version --! 
# !-- Install the below packages using the command: pip install -r requirements.txt --!
# !-- Make sure to install the google python client library using the command: pip install --upgrade google-api-python-client --!
pandas
psycopg2-binary
sqlalchemy
python-dotenv
isodate